requirements.txt
pydantic>=2.5.0,<3.0.0
pydantic-ai>=0.0.12
python-dotenv>=1.0.0
httpx>=0.27.2
openai>=1.12.0
loguru>=0.7.2
typing-extensions>=4.8.0
pytest>=7.4.3
black>=23.11.0
isort>=5.12.0